+++ jupiter-spring-support/src/main/java/org/jupiter/spring/support/JupiterSpringServer.java
@@ -44,10 +44,10 @@ public class JupiterSpringServer implements InitializingBean {
     private RegistryService.RegistryType registryType;
     private JAcceptor acceptor;
 
-    private String registryServerAddresses;             // 注册中心地址 [host1:port1,host2:port2....]
-    private boolean hasRegistryServer;                  // true: 需要连接注册中心; false: IP直连方式
-    private ProviderInterceptor[] providerInterceptors; // 全局拦截器
-    private FlowController<JRequest> flowController;    // 全局流量控制
+    private String registryServerAddresses;                     // 注册中心地址 [host1:port1,host2:port2....]
+    private boolean hasRegistryServer;                          // true: 需要连接注册中心; false: IP直连方式
+    private ProviderInterceptor[] globalProviderInterceptors;   // 全局拦截器
+    private FlowController<JRequest> flowController;            // 全局流量控制
 
     @Override
     public void afterPropertiesSet() throws Exception {
@@ -68,8 +68,8 @@ public class JupiterSpringServer implements InitializingBean {
         }
 
         // 全局拦截器
-        if (providerInterceptors != null && providerInterceptors.length > 0) {
-            server.withGlobalInterceptors(providerInterceptors);
+        if (globalProviderInterceptors != null && globalProviderInterceptors.length > 0) {
+            server.withGlobalInterceptors(globalProviderInterceptors);
         }
 
         // 全局限流
@@ -130,12 +130,12 @@ public class JupiterSpringServer implements InitializingBean {
         this.hasRegistryServer = hasRegistryServer;
     }
 
-    public ProviderInterceptor[] getProviderInterceptors() {
-        return providerInterceptors;
+    public ProviderInterceptor[] getGlobalProviderInterceptors() {
+        return globalProviderInterceptors;
     }
 
-    public void setProviderInterceptors(ProviderInterceptor[] providerInterceptors) {
-        this.providerInterceptors = providerInterceptors;
+    public void setGlobalProviderInterceptors(ProviderInterceptor[] globalProviderInterceptors) {
+        this.globalProviderInterceptors = globalProviderInterceptors;
     }
 
     public FlowController<JRequest> getFlowController() {